correct listing fee calculation

+++ app/helpers/commissions_helper.rb
@@ -37,7 +37,7 @@ module CommissionsHelper
   end
   
   def rounded number
-    number = number.to_f
+    number = number.to_d
     cents = number.round(2).to_s[-2..-1]
     if cents == '00' || cents == '.0'
       number.round
